% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strops.R
\name{:}
\alias{:}
\alias{:.character}
\alias{:.default}
\title{Fill in alphabet letters between two characters.}
\usage{
":"(e1, e2)

\method{:}{character}(e1, e2)

\method{:}{default}(e1, e2)
}
\arguments{
\item{e1}{First character.}

\item{e2}{Second character.}
}
\value{
String with filled in alphabet range.
}
\description{
For the original man page ?`:` redirects to, see '?base::`:`'.
}
\examples{
"a":"c"
# [1] "abc"

}
