% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strops.R
\name{*}
\alias{*}
\alias{*.default}
\alias{*.character}
\title{Replicate string.}
\usage{
"*"(e1, e2)

\method{*}{default}(e1, e2)

\method{*}{character}(e1, e2)
}
\arguments{
\item{e1}{Target string.}

\item{e2}{Number of times to replicate.}
}
\value{
Replicated string.
}
\description{
For the original man page ?`*` redirects to, see '?Arithmetic'.
}
\examples{
"A" * 3
# [1] "AAA"

}
