% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strops.R
\name{^}
\alias{^}
\alias{^.default}
\alias{^.character}
\title{Replace specified characters with spaces.}
\usage{
"^"(e1, e2)

\method{^}{default}(e1, e2)

\method{^}{character}(e1, e2)
}
\arguments{
\item{e1}{Target string.}

\item{e2}{Character(s) to substitute with spaces.}
}
\value{
String e1 with spaces replacing e2 characters.
}
\description{
For the original man page ?`^` redirects to, see '?Arithmetic'.
}
\examples{
"ABC" ^ "B"
# [1] "A C"

}
