% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strops.R
\name{&}
\alias{&}
\alias{&.default}
\alias{&.character}
\title{Check string equality, per character.}
\usage{
"&"(e1, e2)

\method{&}{default}(e1, e2)

\method{&}{character}(e1, e2)
}
\arguments{
\item{e1}{First string.}

\item{e2}{Second string.}
}
\value{
Logical vector.
}
\description{
Check string equality, per character.
}
\examples{
"ABCD" & "ABCE"
# [1] TRUE TRUE TRUE FALSE

}
