% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strops.R
\name{!}
\alias{!}
\alias{!.character}
\alias{!.default}
\title{Switch case.}
\usage{
"!"(x)

\method{!}{character}(x)

\method{!}{default}(x)
}
\arguments{
\item{x}{Target string.}
}
\value{
String with switch case.
}
\description{
For the original man page ?`+` redirects to, see '?Logic'.
}
\examples{
!"abc"
# [1] "ABC"
!"ABC"
# [1] "abc"
!"aBc"
# [1] "AbC"

}
