% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strops.R
\name{+}
\alias{+}
\alias{+.character}
\alias{+.default}
\title{Paste two strings together.}
\usage{
"+"(e1, e2)

\method{+}{character}(e1, e2)

\method{+}{default}(e1, e2)
}
\arguments{
\item{e1}{First string.}

\item{e2}{Second string.}
}
\value{
Pasted string.
}
\description{
For the original man page ?`+` redirects to, see '?Arithmetic'.
}
\examples{
"A" + "B"
# [1] "AB"

}
