% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strops.R
\name{\%..\%}
\alias{\%..\%}
\title{Subset a string.}
\usage{
x \%..\% y
}
\arguments{
\item{x}{Target string.}

\item{y}{Indices to subset on.}
}
\value{
Subsetted string.
}
\description{
Subset a string.
}
\examples{
(~letters) \%..\% c(1:5, 10:15)
# [1] "abcdejklmno"

}
