% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strops.R
\name{~}
\alias{~}
\title{Collapse a character vector.}
\usage{
"~"(x)
}
\arguments{
\item{x}{Character vector to collapse.}
}
\value{
Collapse character vector.
}
\description{
Note: this overrides the original definition for `~`.
To get back the original behaviour, run \code{"~" <- base::`~`}.
}
\examples{
~c("A", "B", "C")
# [1] "ABC"
~1:5
# [1] "12345"

}
