% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{strops-pkg}
\alias{strops-pkg}
\alias{strops-pkg-package}
\title{Operators for handling strings.}
\description{
Operators for handling strings.
}
\details{
Exports methods for common R operators which handle strings.
}
